Move all the GUI related code to gui.py
